requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.6
djangorestframework==3.14.0
drf-spectacular==0.25.1
inflection==0.5.1
jsonschema==4.17.3
kombu==5.2.4
Pillow==9.4.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
pycodestyle==2.10.0
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
uWSGI==2.0.21
vine==5.0.0
wcwidth==0.2.6
flower==1.2.0
requests==2.28.2
pyTelegramBotAPI==4.10.0
django-celery-beat==2.4.0
